pyproject.toml
[project]
name = "python-turtle-art"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Richard Angell", email = "richardangell37@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "beartype>=0.19.0",
    "numpy~=1.26.4",
    "pillow~=10.2.0",
]

[project.scripts]
python-turtle-art = "python_turtle_art.cli:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit==4.1.0",
    "pytest~=8.3",
    "pytest-env==1.1.5",
    "pytest-mock>=3.14.0",
]

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = [
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "E", # pycodestyle
    "F", # Pyflakes
    "I", # isort
    "N", # pep8-naming
    "S", # flake8-bandit
    "SIM", # flake8-simplify
]
ignore = [
    "S311", # suspicious-non-cryptographic-random-usage
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
